# backend/openapi.json
{
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "title": "Detail",
                        "type": "array"
                    }
                },
                "title": "HTTPValidationError",
                "type": "object"
            },
...
            "SpeakingAttemptIn": {
                "properties": {
                    "consent": {
                        "const": true,
                        "title": "Consent",
                        "type": "boolean"
                    },
                    "cue_id": {
                        "title": "Cue Id",
                        "type": "integer"
                    },
                    "duration_ms": {
                        "maximum": 30000.0,
                        "minimum": 1.0,
                        "title": "Duration Ms",
                        "type": "integer"
                    },
                    "self_rating": {
                        "anyOf": [
                            {
                                "enum": [
                                    "repeat",
                                    "almost",
                                    "confident"
                                ],
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Self Rating"
                    }
                },
                "required": [
                    "cue_id",
                    "duration_ms",
                    "consent"
                ],
                "title": "SpeakingAttemptIn",
                "type": "object"
            },
            "SpeakingAttemptOut": {
                "properties": {
                    "consented_at": {
                        "format": "date-time",
                        "title": "Consented At",
                        "type": "string"
                    },
                    "course_slug": {
                        "title": "Course Slug",
                        "type": "string"
                    },
                    "course_title": {
                        "title": "Course Title",
                        "type": "string"
                    },
                    "created_at": {
                        "format": "date-time",
                        "title": "Created At",
                        "type": "string"
                    },
                    "cue_id": {
                        "title": "Cue Id",
                        "type": "integer"
                    },
                    "cue_text": {
                        "title": "Cue Text",
                        "type": "string"
                    },
                    "duration_ms": {
                        "title": "Duration Ms",
                        "type": "integer"
                    },
                    "file_size": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "File Size"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "lesson_number": {
                        "title": "Lesson Number",
                        "type": "integer"
                    },
                    "mime_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mime Type"
                    },
                    "self_rating": {
                        "anyOf": [
                            {
                                "enum": [
                                    "repeat",
                                    "almost",
                                    "confident"
                                ],
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Self Rating"
                    },
                    "status": {
                        "enum": [
                            "pending",
                            "ready"
                        ],
                        "title": "Status",
                        "type": "string"
                    },
                    "transcription": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/TranscriptionJobOut"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "unit_slug": {
                        "title": "Unit Slug",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "course_slug",
                    "course_title",
                    "unit_slug",
                    "lesson_number",
                    "cue_id",
                    "cue_text",
                    "duration_ms",
                    "self_rating",
                    "consented_at",
                    "status",
                    "mime_type",
                    "file_size",
                    "created_at"
                ],
                "title": "SpeakingAttemptOut",
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "Aulas de Inglês — API",
        "version": "0.1.0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/speaking/attempts": {
            "get": {
                "operationId": "listar_tentativas_api_speaking_attempts_get",
                "parameters": [
                    {
                        "in": "query",
                        "name": "lesson",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "minimum": 1,
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Lesson"
                        }
                    },
                    {
                        "in": "query",
                        "name": "course",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "maxLength": 100,
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Course"
                        }
                    },
                    {
                        "in": "query",
                        "name": "unit",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "maxLength": 100,
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Unit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SpeakingAttemptOut"
                                    },
                                    "title": "Response Listar Tentativas Api Speaking Attempts Get",
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Listar Tentativas",
                "tags": [
                    "speaking"
                ]
            },
            "post": {
                "operationId": "criar_tentativa_api_speaking_attempts_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpeakingAttemptIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpeakingAttemptOut"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Criar Tentativa",
                "tags": [
                    "speaking"
                ]
            }
        },
        "/api/speaking/attempts/{attempt_id}": {
            "delete": {
                "operationId": "excluir_tentativa_api_speaking_attempts__attempt_id__delete",
                "parameters": [
                    {
                        "in": "path",
                        "name": "attempt_id",
                        "required": true,
                        "schema": {
                            "title": "Attempt Id",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Excluir Tentativa",
                "tags": [
                    "speaking"
                ]
            }
        },
        "/api/speaking/attempts/{attempt_id}/audio": {
            "get": {
                "operationId": "obter_audio_api_speaking_attempts__attempt_id__audio_get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "attempt_id",
                        "required": true,
                        "schema": {
                            "title": "Attempt Id",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Obter Audio",
                "tags": [
                    "speaking"
                ]
            },
            "put": {
                "operationId": "enviar_audio_api_speaking_attempts__attempt_id__audio_put",
                "parameters": [
                    {
                        "in": "path",
                        "name": "attempt_id",
                        "required": true,
                        "schema": {
                            "title": "Attempt Id",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SpeakingAttemptOut"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Enviar Audio",
                "tags": [
                    "speaking"
                ]
            }
        },
        "/api/speaking/attempts/{attempt_id}/transcription": {
            "post": {
                "operationId": "solicitar_transcricao_api_speaking_attempts__attempt_id__transcription_post",
                "parameters": [
                    {
                        "in": "path",
                        "name": "attempt_id",
                        "required": true,
                        "schema": {
                            "title": "Attempt Id",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranscriptionJobOut"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Solicitar Transcricao",
                "tags": [
                    "speaking"
                ]
            }
        },
        "/api/speaking/transcriptions/{job_id}": {
            "delete": {
                "operationId": "excluir_transcricao_api_speaking_transcriptions__job_id__delete",
                "parameters": [
                    {
                        "in": "path",
                        "name": "job_id",
                        "required": true,
                        "schema": {
                            "title": "Job Id",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Excluir Transcricao",
                "tags": [
                    "speaking"
                ]
            }
        },
        "/api/speaking/transcriptions/{job_id}/rating": {
            "put": {
                "operationId": "avaliar_transcricao_api_speaking_transcriptions__job_id__rating_put",
                "parameters": [
                    {
                        "in": "path",
                        "name": "job_id",
                        "required": true,
                        "schema": {
                            "title": "Job Id",
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HumanRatingIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TranscriptionJobOut"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Avaliar Transcricao",
                "tags": [
                    "speaking"
                ]
            }
        },
...
    }
}